00005-emb-lm-skip: fix 3111 by turning compressor to bfloat16

+++ experiments/00005-emb-lm-skip/runs/3111-concat-skip11-last-mlp-with-projection.py
@@ -231,7 +231,7 @@ class GPT(nn.Module):
         self.value_embeds = nn.ModuleList([nn.Embedding(vocab_size, model_dim) for _ in range(5)])
         self.blocks = nn.ModuleList([Block(model_dim, num_heads, max_seq_len, i) for i in range(num_layers)])
         self.out_mlp = MLP(dim=model_dim+compressed_dim, hdim=model_dim+compressed_dim)
-        self.compressor = nn.Parameter(init_linear(torch.empty(compressed_dim, model_dim)))
+        self.compressor = nn.Parameter(init_linear(torch.empty(compressed_dim, model_dim)).bfloat16())
         # there are only 50257 unique GPT-2 tokens; we extend to nearest multiple of 128 for efficiency.
         # suggested to me by @Grad62304977. this originates from Karpathy's experiments.
         self.lm_head_w = nn.Parameter(torch.zeros(next_multiple_of_n(vocab_size, n=128), model_dim + compressed_dim))
